// Servers/models/ISO-42001/subClauseStructISO.model.ts
import { Column, DataType, ForeignKey, Model, Table } from "sequelize-typescript";
import { ClauseStructISOModel } from "./clauseStructISO.model";

export type SubClauseStructISO = {
  id?: number; // auto generated by database
  title: string;
  order_no: number;
  summary: string;
  questions: string[];
  evidence_examples: string[];
  clause_id?: number;
}

@Table({
  tableName: "subclause_struct_iso"
})
export class SubClauseStructISOModel extends Model<SubClauseStructISO> {
  @Column({
    type: DataType.INTEGER,
    autoIncrement: true,
    primaryKey: true,
  })
  id!: number;

  @Column({
    type: DataType.STRING
  })
  title!: string;

  @Column({
    type: DataType.INTEGER
  })
  order_no!: number;

  @Column({
    type: DataType.STRING
  })
  summary!: string;

  @Column({
    type: DataType.ARRAY(DataType.STRING)
  })
  questions!: string[];

  @Column({
    type: DataType.ARRAY(DataType.STRING)
  })
  evidence_examples!: string[];

  @ForeignKey(() => ClauseStructISOModel)
  @Column({
    type: DataType.INTEGER
  })
  clause_id!: number;
}